Fix API endpoint issues for Next.js 16
- Fixed params handling in [id] route (now awaits Promise<{id}>)
- Fixed tools_count to return number instead of string (::int cast)

All API endpoints now working correctly:
- POST /api/registry/register ✅
- GET /api/registry/servers ✅
- GET /api/registry/servers/:id ✅

+++ src/app/api/registry/servers/route.ts
@@ -35,7 +35,7 @@ export async function GET(request: NextRequest) {
         version: mcpServers.version,
         last_health_check: mcpServers.lastHealthCheck,
         tools_count: sql<number>`(
-          SELECT COUNT(*)
+          SELECT COUNT(*)::int
           FROM ${tools}
           WHERE ${tools.serverId} = ${mcpServers.id}
         )`,
